pyproject.toml
[project]
name = "expense-analyzer-backend"
version = "0.1.0"
description = "Smart expense analyzer API for personal finance tracking"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "flask>=2.3.3",
    "flask-cors>=4.0.0",
    "pymongo>=4.5.0",
    "pandas>=2.1.1",
    "pyjwt>=2.8.0",
    "python-dotenv>=1.0.0",
    "marshmallow>=3.20.1",
    "bcrypt>=4.0.1",
    "python-dateutil>=2.8.2",
    "gunicorn>=21.2.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-flask>=1.2.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "pre-commit>=3.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["expense_analyzer_backend"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"